Finder Test
: Write a program which accepts two inputs, representing two playing
cards out of a standard 52 card deck. Add these two cards together to
produce a blackjack score, and print the score on the screen for the
output.
Cards will be identified by the input, the first part representing the
face value from 2-10, plus A, K, Q, J. The second part represents the
suit S, C, D, H.
The blackjack score is the face value of the two cards added together,
with cards 2-10 being the numeric face value, and A is worth 11, and K,
Q, J are each worth 10.